pyproject.toml
[project]
name = "hpc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "John Bradley", email = "johnbradley2008@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "gradio",
    "python-dotenv",
    "anthropic",
]

[project.scripts]
hpc-agent = "hpca.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hpca"]
